Request beacons
