+++ state_machine/applications/flight/lib/radio_utils/headers.py
@@ -9,3 +9,5 @@ DISK_BUFFERED_MID = 0xfb
 DISK_BUFFERED_END = 0xfa
 
 COMMAND = 0x01
+
+BEACON = 0x02
